switch cli to kind verb noun
